+++ ktrain/text/preprocessor.py
@@ -163,6 +163,8 @@ def bert_tokenize(docs, tokenizer, max_length, verbose=1):
     indices = []
     for i in mb:
         for doc in pb:
+            # https://stackoverflow.com/questions/67360987/bert-model-bug-encountered-during-training/67375675#67375675
+            doc = str(doc) if isinstance(doc, (float, int)) else doc
             ids, segments = tokenizer.encode(doc, max_len=max_length)
             indices.append(ids)
         if verbose: mb.write('done.')